Refactor: 시간 바꿔주는 함수 분리
